rename rotationTests package to just rotation, collapse some other rotation tests into 1 file.
