--- EduBot/tests/xbot/edubot/rotationTests/GoLeft90FromNeg150Test.java
@@ -1,13 +1,0 @@
-package xbot.edubot.rotationTests;
-
-import org.junit.Test;
-
-public class GoLeft90FromNeg150Test extends BaseOrientationEngineTest {
-	
-	@Test
-	public void goLeft90FromNeg150()
-	{
-		run90ClassTest(-150, -60);
-	}
-
-}
