--- EduBot/tests/xbot/edubot/rotationTests/GoLeft90FromStartTest.java
@@ -1,11 +1,0 @@
-package xbot.edubot.rotationTests;
-
-import org.junit.Test;
-
-public class GoLeft90FromStartTest extends BaseOrientationEngineTest {
-	@Test
-	public void goLeft90FromStart()
-	{
-		run90ClassTest(0, 90);
-	}
-}
